+++ R/valytics-package.R
@@ -2,7 +2,12 @@
 "_PACKAGE"
 
 ## usethis namespace: start
-#' @importFrom ggplot2 .data
+#' @importFrom ggplot2 .data autoplot
 #' @importFrom stats complete.cases median qnorm qt sd shapiro.test
+#' @importFrom utils modifyList
 ## usethis namespace: end
 NULL
+
+#' @importFrom ggplot2 autoplot
+#' @export
+ggplot2::autoplot